Wakame

Documentation for Wakame.

Features

Wakame is a powerful and versatile library that can be used to segment a wide variety of Japanese text corpora. It is also highly efficient, making it ideal for large-scale text processing tasks.

Here are some of the key features of Wakame:

Port of MeCab: Wakame is a port of the popular Japanese text segmentation library, MeCab. This gives Wakame access to MeCab's vast vocabulary and morphological analysis capabilities.

Interface with C++: Wakame is written in Julia, but it interfaces directly with C++ code. This allows Wakame to take advantage of the speed and efficiency of C++ while still providing the flexibility and expressiveness of Julia.

Support for a variety of corpora: Wakame supports a variety of Japanese text corpora, including news articles, social media posts, and academic papers.

High performance: Wakame is highly efficient, making it ideal for large-scale text processing tasks.

Installation

To install Wakame either do

using Pkg
Pkg.add("Wakame")

or switch to Pkg mode with ] and issue

pkg> add Wakame

Requirement

This package requires the following software to be installed:

mecab

dictionary for mecab (such as mecab-ipadic, mecab-naist-jdic, and others)

Usage

using Wakame

# Create a mecab tagger
mecab = Mecab()

# You can give MeCab option like "-o wakati"
# mecab = Mecab("-o wakati")

# Parse text. It returns Array of MecabNode type
results = parse(mecab, "すももももももももものうち")

# Access each result. It returns Array of String
for result in results
  println(result.surface, ":", result.feature)
end

# Parse surface
results = parse_surface(mecab, "すももももももももものうち")

# Access each result. It returns Array of Array of MecabNode
for result in results
  println(result)
end

# Parse nbest result
nbest_results = parse_nbest(mecab, 3, "すももももももももものうち")
for nbest_result in nbest_results
  for result in nbest_result
    println(result.surface, ":", result.feature)
  end
  println()
end

Credits

Reboot of MeCab.jl created by Michiaki Ariga

Original MeCab is created by Taku Kudo

Contributors

r9y9

snthot

tkelman

Wakame.Mecab

Wakame.MecabNode

Wakame.MecabRawNode

Base.parse

Wakame.create_mecab_results

Wakame.create_node

Wakame.create_nodes

Wakame.create_surface

Wakame.create_surfaces

Wakame.mecab_result

Wakame.mecab_sparse_tonode

Wakame.nbest_init

Wakame.nbest_next_tostr

Wakame.nbest_sparse_tostr

Wakame.parse_nbest

Wakame.parse_surface

Wakame.parse_surface2

Wakame.sparse_tostr

Wakame.Mecab — Type

Mecab

A mutable struct representing a MeCab tagger. It is created by calling the Mecab() constructor.

Examples

tagger = Mecab()
results = parse(mecab, "すももももももももものうち")

Wakame.MecabNode — Type

A mutable struct representing a MeCab node.

Wakame.MecabRawNode — Type

A mutable struct representing a raw node in MeCab used to create a `MecabNode`

Base.parse — Method

Parse a string using a `Mecab` tagger and return result as an `Array{MecabNode}`

Wakame.create_mecab_results — Method

Create an array of `MecabNode` from an array of MeCab results

Wakame.create_node — Method

Function to create a `MecabNode` from a `MecabRawNode`

Wakame.create_nodes — Method

Function to create an `Array{MecabNode}` from a `Ptr{MecabRawNode}`

Wakame.create_surface — Method

Function to create a `String` from a `MecabRawNode`

Wakame.create_surfaces — Method

Function to create an `Array{String}` from a `Ptr{MecabRawNode}`

Wakame.mecab_result — Method

Create a `MecabNode` from a MeCab result

Wakame.mecab_sparse_tonode — Method

Parse a string using a `Mecab` tagger and return result as a `Ptr{MecabRawNode}`

Wakame.nbest_init — Method

Initialize the n-best enumeration process for a `Mecab` tagger

Wakame.nbest_next_tostr — Method

Get the next n-best result from a `Mecab` tagger

Wakame.nbest_sparse_tostr — Method

Parse a string using a `Mecab` tagger and return top n results as a string

Wakame.parse_nbest — Method

Parse a string using a `Mecab` tagger and return the top n results as an `Array{Array{MecabNode}}`

Wakame.parse_surface — Method

Parse a string using a `Mecab` tagger and return surface strings as an `Array{String}`

Wakame.parse_surface2 — Method

Parse a string using a `Mecab` tagger and return surface strings as an `Array{String}`

Wakame.sparse_tostr — Method

Parse a string using a `Mecab` tagger and return result as a string